+++ pyodesys/_release.py
@@ -1,1 +1,2 @@
-__version__ = '0.4.0.git'
+__version__ = '0.4.0.dev0'
+# see https://www.python.org/dev/peps/pep-0440/