+++ pyodesys/symbolic.py
@@ -3,9 +3,9 @@
 from __future__ import absolute_import, division, print_function
 
 from itertools import chain, repeat
+import os
 
 import numpy as np
-import sympy as sp
 
 from .core import OdeSys
 from .util import (
@@ -14,61 +14,180 @@ from .util import (
 )
 
 
+def _lambdify(backend=None):
+    backend = os.environ.get('PYODESYS_SYM_BACKEND', 'sympy')
+    if backend == 'sympy':
+        import sympy as sp
+
+        def lambdify(*args, **kwargs):
+            if 'modules' not in kwargs:
+                kwargs['modules'] = [{'ImmutableMatrix': np.array}, 'numpy']
+            return sp.lambdify(*args, **kwargs)
+        return lambdify
+    elif backend == 'pysym':
+        import pysym as ps
+        return ps.Lambdify
+    else:
+        raise NotImplementedError('Unknown symbolic package: %s' %
+                                  backend)
+
+
+def _lambdify_unpack(backend=None):
+    backend = os.environ.get('PYODESYS_SYM_BACKEND', 'sympy')
+    if backend == 'sympy':
+        return True
+    elif backend == 'pysym':
+        return False
+    else:
+        raise NotImplementedError('Unknown symbolic package: %s' %
+                                  backend)
+
+
+def _Matrix(backend=None):
+    backend = os.environ.get('PYODESYS_SYM_BACKEND', 'sympy')
+    if backend == 'sympy':
+        import sympy as sp
+        return sp.Matrix
+    elif backend == 'pysym':
+        import pysym as ps
+        return ps.Matrix
+    else:
+        raise NotImplementedError('Unknown symbolic package: %s' %
+                                  backend)
+
+
+def _Symbol(backend=None):
+    backend = os.environ.get('PYODESYS_SYM_BACKEND', 'sympy')
+    if backend == 'sympy':
+        import sympy as sp
+
+        def Symbol(*args, **kwargs):
+            if 'real' not in kwargs:
+                kwargs['real'] = True
+            return sp.Symbol(*args, **kwargs)
+        return Symbol
+    elif backend == 'pysym':
+        import pysym as ps
+        return ps.Symbol
+    else:
+        raise NotImplementedError('Unknown symbolic package: %s' %
+                                  backend)
+
+
+def _Dummy(backend=None):
+    backend = os.environ.get('PYODESYS_SYM_BACKEND', 'sympy')
+    if backend == 'sympy':
+        import sympy as sp
+        return sp.Dummy
+    elif backend == 'pysym':
+        import pysym as ps
+        return ps.Dummy
+    else:
+        raise NotImplementedError('Unknown symbolic package: %s' %
+                                  backend)
+
+
+def _symarray(backend=None):
+    backend = os.environ.get('PYODESYS_SYM_BACKEND', 'sympy')
+    if backend == 'sympy':
+        import sympy as sp
+
+        def symarray(prefix, shape, Symbol=None, real=True):
+            """ Creates an nd-array of symbols
+
+            Parameters
+            ----------
+            prefix: str
+            shape: tuple
+            Symbol: callable
+                (defualt :func:`Symbol`)
+            """
+            # see https://github.com/sympy/sympy/pull/9939
+            # when released: return sp.symarray(key, n, real=True)
+            arr = np.empty(shape, dtype=object)
+            for index in np.ndindex(shape):
+                arr[index] = (
+                    Symbol or (lambda name: sp.Symbol(name, real=real))
+                )('%s_%s' % (prefix, '_'.join(map(str, index))))
+            return arr
+        return symarray
+    elif backend == 'pysym':
+        import pysym as ps
+        return ps.symarray
+    else:
+        raise NotImplementedError('Unknown symbolic package: %s' %
+                                  backend)
+
+
 class SymbolicSys(OdeSys):
     """
+    ODE System from symbolic expressions
+
+    Creates a :class:`OdeSys` instance
+    from symbolic expressions. Jacboian and second derivatives
+    are derived when needed.
+
     Parameters
     ----------
     dep_exprs: iterable of (symbol, expression)-pairs
     indep: symbol
         independent variable (default: None => autonomous system)
     jac: ImmutableMatrix or bool (default: True)
-        If True:
+        if True:
             calculate jacobian from exprs
-        If False:
+        if False:
             do not compute jacobian (use explicit steppers)
-        If ImmutableMatrix:
+        if instance of ImmutableMatrix:
             user provided expressions for the jacobian
     roots: iterable of expressions
         equations to look for root's for during integration
         (currently available through cvode)
     lambdify: callback
-        default: sympy.lambdify
+        default: :py:func:`sympy.lambdify`
     lambdify_unpack: bool (default: True)
         whether or not unpacking of args needed when calling lambdify callback
     Matrix: class
-        default: sympy.Matrix
+        default: :py:class:`sympy.Matrix`
+    Symbol: class
+        default: :py:class:`sympy.Symbol`
+    Dummy: class
+        default: :py:class:`sympy.Dummy`
+    symarray: callback
+        default: :py:class:`sympy.symarray`
     \*\*kwargs:
         See :py:class:`OdeSys`
 
     Attributes
     ----------
-    dep : iterable of dependent variables
-    indep : independent variable
-    params : iterable of parameters
+    dep : iterable of symbols
+        dependent variables
+    indep : symbol
+        independent variable
+    params : iterable of symbols
+        parameters
 
     Notes
     -----
-    Works for a moderate number of unknowns, sympy.lambdify has
+    Works for a moderate number of unknowns, :py:func:`sympy.lambdify` has
     an upper limit on number of arguments.
     """
 
     def __init__(self, dep_exprs, indep=None, params=(), jac=True, dfdx=True,
-                 roots=None, lambdify=None, lambdify_unpack=True, Matrix=None,
-                 **kwargs):
+                 roots=None, lambdify=None, lambdify_unpack=None, Matrix=None,
+                 Symbol=None, Dummy=None, symarray=None, **kwargs):
         self.dep, self.exprs = zip(*dep_exprs)
         self.indep = indep
         self.params = params
         self._jac = jac
         self._dfdx = dfdx
         self.roots = roots
-        if lambdify is not None:
-            self.lambdify = lambdify
-        self.lambdify_unpack = lambdify_unpack
-        if Matrix is None:
-            import sympy
-            self.Matrix = sympy.ImmutableMatrix
-        else:
-            self.Matrix = Matrix
+        self.lambdify = lambdify or _lambdify()
+        self.lambdify_unpack = (_lambdify_unpack() if lambdify_unpack is None
+                                else lambdify_unpack)
+        self.Matrix = Matrix or _Matrix()
+        self.Symbol = Symbol or _Symbol()
+        self.Dummy = Dummy or _Dummy()
+        self.symarray = symarray or _symarray()
         # we need self.band before super().__init__
         self.band = kwargs.get('band', None)
         if kwargs.get('names', None) is True:
@@ -81,44 +200,39 @@ class SymbolicSys(OdeSys):
             nroots=None if roots is None else len(roots),
             **kwargs)
 
-    @staticmethod
-    def Symbol(name):
-        return sp.Symbol(name, real=True)
-
     @classmethod
-    def symarray(cls, prefix, shape, Symbol=None):
-        # see https://github.com/sympy/sympy/pull/9939
-        # when released: return sp.symarray(key, n, real=True)
-        arr = np.empty(shape, dtype=object)
-        for index in np.ndindex(shape):
-            arr[index] = (Symbol or cls.Symbol)('%s_%s' % (
-                prefix, '_'.join(map(str, index))))
-        return arr
+    def from_callback(cls, cb, ny, nparams=0, *args, **kwargs):
+        """ Create an instance from a callback.
 
-    @staticmethod
-    def lambdify(*args, **kwargs):
-        if 'modules' not in kwargs:
-            kwargs['modules'] = [{'ImmutableMatrix': np.array}, 'numpy']
-        return sp.lambdify(*args, **kwargs)
+        Parameters
+        ----------
+        cb: callbable
+            Signature rhs(x, y[:], p[:]) -> f[:]
+        ny: int
+            length of y
+        nparams: int (default: 0)
+            length of p
+        \*args:
+            arguments passed onto :class:`SymbolicSys`
+        \*\*kwargs:
+            keyword arguments passed onto :class:`SymbolicSys`
 
-    @classmethod
-    def num_transformer_factory(cls, fw, bw, inp, lambdify=None):
-        lambdify = lambdify or cls.lambdify
-        return lambdify(inp, fw), lambdify(inp, bw)
-
-    @classmethod
-    def from_callback(cls, cb, n, nparams=0, *args, **kwargs):
-        x = cls.Symbol('x')
-        y = cls.symarray('y', n)
-        p = cls.symarray('p', nparams)
+        Returns
+        -------
+        An instance of :class:`SymbolicSys`.
+        """
+        x = kwargs.get('Symbol', _Symbol())('x')
+        y = kwargs.get('symarray', _symarray())('y', ny)
+        p = kwargs.get('symarray', _symarray())('p', nparams)
         exprs = ensure_3args(cb)(x, y, p)
         return cls(zip(y, exprs), x, p, *args, **kwargs)
 
     @property
     def ny(self):
+        """ Number of dependent variables in the system. """
         return len(self.exprs)
 
-    def args(self, x=None, y=None, params=()):
+    def _args(self, x=None, y=None, params=()):
         if x is None:
             x = self.indep
         if y is None:
@@ -129,6 +243,7 @@ class SymbolicSys(OdeSys):
         return args + tuple(params)
 
     def get_jac(self):
+        """ Derive the jacobian from ``self.exprs`` and ``self.dep``. """
         if self._jac is True:
             if self.band is None:
                 f = self.Matrix(1, self.ny, lambda _, q: self.exprs[q])
@@ -143,6 +258,7 @@ class SymbolicSys(OdeSys):
         return self._jac
 
     def get_dfdx(self):
+        """ Calculate seond order derivatives in independent variable """
         if self._dfdx is True:
             if self.indep is None:
                 self._dfdx = [0]*self.ny
@@ -153,51 +269,55 @@ class SymbolicSys(OdeSys):
         return self._dfdx
 
     def get_f_ty_callback(self):
-        cb = self.lambdify(list(chain(self.args(), self.params)), self.exprs)
+        """ Generate a callback for evaluating ``self.exprs``. """
+        cb = self.lambdify(list(chain(self._args(), self.params)), self.exprs)
 
         def f(x, y, params=()):
             if self.lambdify_unpack:
-                return np.asarray(cb(*self.args(x, y, params)))
+                return np.asarray(cb(*self._args(x, y, params)))
             else:
-                return np.asarray(cb(self.args(x, y, params)))
+                return np.asarray(cb(self._args(x, y, params)))
         return f
 
     def get_j_ty_callback(self):
+        """ Generate a callback for evaluating the jacobian. """
         j_exprs = self.get_jac()
         if j_exprs is False:
             return None
-        cb = self.lambdify(list(chain(self.args(), self.params)), j_exprs)
+        cb = self.lambdify(list(chain(self._args(), self.params)), j_exprs)
 
         def j(x, y, params=()):
             if self.lambdify_unpack:
-                return np.asarray(cb(*self.args(x, y, params)))
+                return np.asarray(cb(*self._args(x, y, params)))
             else:
-                return np.asarray(cb(self.args(x, y, params)))
+                return np.asarray(cb(self._args(x, y, params)))
         return j
 
     def get_dfdx_callback(self):
+        """ Generate a callback for evaluating derivative of ``self.exprs`` """
         dfdx_exprs = self.get_dfdx()
         if dfdx_exprs is False:
             return None
-        cb = self.lambdify(list(chain(self.args(), self.params)), dfdx_exprs)
+        cb = self.lambdify(list(chain(self._args(), self.params)), dfdx_exprs)
 
         def dfdx(x, y, params=()):
             if self.lambdify_unpack:
-                return np.asarray(cb(*self.args(x, y, params)))
+                return np.asarray(cb(*self._args(x, y, params)))
             else:
-                return np.asarray(cb(self.args(x, y, params)))
+                return np.asarray(cb(self._args(x, y, params)))
         return dfdx
 
     def get_roots_callback(self):
+        """ Generate a callback for evaluating ``self.roots`` """
         if self.roots is None:
             return None
-        cb = self.lambdify(list(chain(self.args(), self.params)), self.roots)
+        cb = self.lambdify(list(chain(self._args(), self.params)), self.roots)
 
         def roots(x, y, params=()):
             if self.lambdify_unpack:
-                return np.asarray(cb(*self.args(x, y, params)))
+                return np.asarray(cb(*self._args(x, y, params)))
             else:
-                return np.asarray(cb(self.args(x, y, params)))
+                return np.asarray(cb(self._args(x, y, params)))
         return roots
 
     # Not working yet:
@@ -228,6 +348,27 @@ class SymbolicSys(OdeSys):
 
 
 class TransformedSys(SymbolicSys):
+    """ SymbolicSys with abstracted variable transformations.
+
+    Parameters
+    ----------
+    dep_exprs: iterable of pairs
+        see :class:`SymbolicSys`
+    indep: Symbol
+        see :class:`SymbolicSys`
+    dep_transf: iterable of (expression, expression) pairs
+        pairs of (forward, backward) transformations for the dependents
+        variables
+    indep_transf: pair of expressions
+        forward and backward transformation of the independent variable
+    params: see :class:`SymbolicSys`
+    exprs_process_cb: callbable
+        signatrue f(exprs) -> exprs
+        post processing of the expressions for the derivatives of the
+        dependent variables after transformation have been applied.
+    \*\*kwargs:
+        keyword arguments passed onto :class:`SymbolicSys`
+    """
 
     def __init__(self, dep_exprs, indep=None, dep_transf=None,
                  indep_transf=None, params=(), exprs_process_cb=None,
@@ -252,25 +393,45 @@ class TransformedSys(SymbolicSys):
 
         super(TransformedSys, self).__init__(
             zip(dep, exprs), indep, params,
-            pre_processors=[self.forward_transform_xy],
-            post_processors=[self.back_transform_out], **kwargs)
+            pre_processors=[self._forward_transform_xy],
+            post_processors=[self._back_transform_out], **kwargs)
         # the pre- and post-processors need callbacks:
-        args = self.args(indep, dep, params)
-        self.f_dep, self.b_dep = self.num_transformer_factory(
-            self.dep_fw, self.dep_bw, args)
+        args = self._args(indep, dep, params)
+        self.f_dep = self.lambdify(args, self.dep_fw)
+        self.b_dep = self.lambdify(args, self.dep_bw)
         if (self.indep_fw, self.indep_bw) != (None, None):
-            self.f_indep, self.b_indep = self.num_transformer_factory(
-                self.indep_fw, self.indep_bw, args)
+            self.f_indep = self.lambdify(args, self.indep_fw)
+            self.b_indep = self.lambdify(args, self.indep_bw)
         else:
             self.f_indep = None
             self.b_indep = None
 
     @classmethod
-    def from_callback(cls, cb, n, nparams=0, dep_transf_cbs=None,
+    def from_callback(cls, cb, ny, nparams=0, dep_transf_cbs=None,
                       indep_transf_cbs=None, **kwargs):
-        x = cls.Symbol('x')
-        y = cls.symarray('y', n)
-        p = cls.symarray('p', nparams)
+        """
+        Create an instance from a callback.
+
+        Analogous to :func:`SymbolicSys.from_callback`.
+
+        Parameters
+        ----------
+        cb: callable
+            Signature rhs(x, y[:], p[:]) -> f[:]
+        ny: int
+            length of y
+        nparams: int
+            length of p
+        dep_transf_cbs: iterable of pairs callables
+            callables should have the signature f(yi) -> expression in yi
+        indep_transf_cbs: pair of callbacks
+            callables should have the signature f(x) -> expression in x
+        \*\*kwargs:
+            keyword arguments passed onto :class:`TransformedSys`
+        """
+        x = kwargs.get('Symbol', _Symbol())('x')
+        y = kwargs.get('symarray', _symarray())('y', ny)
+        p = kwargs.get('symarray', _symarray())('p', nparams)
         exprs = ensure_3args(cb)(x, y, p)
         if dep_transf_cbs is not None:
             dep_transf = [(fw(yi), bw(yi)) for (fw, bw), yi
@@ -286,8 +447,8 @@ class TransformedSys(SymbolicSys):
         return cls(list(zip(y, exprs)), x, dep_transf,
                    indep_transf, p, **kwargs)
 
-    def back_transform_out(self, xout, yout, params):
-        args = self.args(xout, yout.T, params)
+    def _back_transform_out(self, xout, yout, params):
+        args = self._args(xout, yout.T, params)
         if self.lambdify_unpack:
             return (xout if self.b_indep is None else self.b_indep(*args),
                     np.array(self.b_dep(*args)).T, params)
@@ -295,8 +456,8 @@ class TransformedSys(SymbolicSys):
             return (xout if self.b_indep is None else self.b_indep(args),
                     np.array(self.b_dep(args)).T, params)
 
-    def forward_transform_xy(self, x, y, p):
-        args = self.args(x, y, p)
+    def _forward_transform_xy(self, x, y, p):
+        args = self._args(x, y, p)
         if self.lambdify_unpack:
             return (x if self.f_indep is None else self.f_indep(*args),
                     self.f_dep(*args), p)
@@ -317,7 +478,12 @@ def symmetricsys(dep_tr=None, indep_tr=None, **kwargs):
     indep_tr: pair of callables (default: None)
         Forward and backward transformation to be applied to the
         independent variable
+    \*\*kwargs:
+        default keyword arguments for the TransformedSys subclass
 
+    Returns
+    -------
+    Subclass of TransformedSys
 
     Examples
     --------
@@ -351,11 +517,11 @@ def symmetricsys(dep_tr=None, indep_tr=None, **kwargs):
                 **new_kwargs)
 
         @classmethod
-        def from_callback(cls, cb, n, nparams=0, **inner_kwargs):
+        def from_callback(cls, cb, ny, nparams=0, **inner_kwargs):
             new_kwargs = kwargs.copy()
             new_kwargs.update(inner_kwargs)
             return TransformedSys.from_callback(
-                cb, n, nparams,
+                cb, ny, nparams,
                 dep_transf_cbs=repeat(dep_tr) if dep_tr is not None else None,
                 indep_transf_cbs=indep_tr,
                 **new_kwargs)
@@ -363,9 +529,25 @@ def symmetricsys(dep_tr=None, indep_tr=None, **kwargs):
 
 
 class ScaledSys(TransformedSys):
+    """
+    Special case of TransformedSys where the variables have been scaled.
+
+    Parameters
+    ----------
+    dep_exprs: see :class:`SymbolicSys`
+    indep: see :class:`SymbolicSys`
+    dep_scaling: number (>0) or iterable of numbers
+        scaling of the dependent variables (default: 1)
+    indep_scaling: number (>0)
+        scaling of the independent variable (default: 1)
+    params: see :class:`SymbolicSys`
+    \*\*kwargs:
+        keyword arguments passed onto TransformedSys
+
+    """
 
     @staticmethod
-    def scale_fw_bw(scaling):
+    def _scale_fw_bw(scaling):
         return (lambda x: scaling*x, lambda x: x/scaling)
 
     def __init__(self, dep_exprs, indep=None, dep_scaling=1, indep_scaling=1,
@@ -376,8 +558,8 @@ class ScaledSys(TransformedSys):
         except TypeError:
             n = len(dep_exprs)
             dep_scaling = [dep_scaling]*n
-        transf_dep_cbs = [self.scale_fw_bw(s) for s in dep_scaling]
-        transf_indep_cbs = self.scale_fw_bw(indep_scaling)
+        transf_dep_cbs = [self._scale_fw_bw(s) for s in dep_scaling]
+        transf_indep_cbs = self._scale_fw_bw(indep_scaling)
         super(ScaledSys, self).__init__(
             dep_exprs, indep,
             dep_transf=[(transf_cb[0](depi),
@@ -388,18 +570,33 @@ class ScaledSys(TransformedSys):
             None, **kwargs)
 
     @classmethod
-    def from_callback(cls, cb, n, nparams=0, dep_scaling=1, indep_scaling=1,
+    def from_callback(cls, cb, ny, nparams=0, dep_scaling=1, indep_scaling=1,
                       **kwargs):
+        """
+        Create an instance from a callback.
+
+        Analogous to :func:`SymbolicSys.from_callback`.
+
+        Parameters
+        ----------
+        cb: callable
+            Signature rhs(x, y[:], p[:]) -> f[:]
+        ny: int
+            length of y
+        nparams: int
+            length of p
+        dep_scaling: number (>0) or iterable of numbers
+            scaling of the dependent variables (default: 1)
+        indep_scaling: number (>0)
+            scaling of the independent variable (default: 1)
+        \*\*kwargs:
+            keyword arguments passed onto :class:`ScaledSys`
+        """
         return TransformedSys.from_callback(
-            cb, n, nparams,
-            dep_transf_cbs=repeat(cls.scale_fw_bw(dep_scaling)),
-            indep_transf_cbs=cls.scale_fw_bw(indep_scaling),
+            cb, ny, nparams,
+            dep_transf_cbs=repeat(cls._scale_fw_bw(dep_scaling)),
+            indep_transf_cbs=cls._scale_fw_bw(indep_scaling),
         )
-
-
-def _take(indices, iterable):
-    return np.asarray([elem for idx, elem in enumerate(
-        iterable) if idx in indices])
 
 
 def _skip(indices, iterable):
@@ -429,9 +626,12 @@ class PartiallySolvedSystem(SymbolicSys):
     analytic_factory: callable
         signature: solved(x0, y0, p0) -> dict, where dict maps
         independent variables as analytic expressions in remaining variables
+    Dummy: callable (default: None)
+        if None use self.Dummy
 
     Examples
     --------
+    >>> import sympy as sp
     >>> odesys = SymbolicSys.from_callback(
     ...     lambda x, y, p: [
     ...         -p[0]*y[0],
@@ -456,8 +656,7 @@ class PartiallySolvedSystem(SymbolicSys):
         self.analytic_factory = analytic_factory
         if original_system.roots is not None:
             raise NotImplementedError('roots unsupported')
-        if Dummy is None:
-            Dummy = sp.Dummy
+        Dummy = Dummy or _Dummy()
         self.init_indep = Dummy()
         self.init_dep = [Dummy() for _ in range(original_system.ny)]
 
@@ -507,7 +706,10 @@ class PartiallySolvedSystem(SymbolicSys):
             zip(new_dep, new_exprs), original_system.indep, new_params,
             lambdify=original_system.lambdify,
             lambdify_unpack=original_system.lambdify_unpack,
-            Matrix=original_system.Matrix, **new_kw)
+            Matrix=original_system.Matrix,
+            Symbol=original_system.Symbol,
+            Dummy=original_system.Dummy,
+            **new_kw)
 
     def _get_analytic_cb(self, ori_sys, analytic_exprs, new_params):
         cb = ori_sys.lambdify(_concat(ori_sys.indep, new_params),