+++ pyodesys/util.py
@@ -7,16 +7,36 @@ import inspect
 
 
 def stack_1d_on_left(x, y):
+    """ Stack a 1D array on the left side of a 2D array
+
+    Parameters
+    ----------
+    x: 1D array
+    y: 2D array
+        Requirement: ``shape[0] == x.size``
+    """
     return np.hstack((np.asarray(x).reshape(len(x), 1),
                       np.asarray(y)))
 
 
 def banded_jacobian(y, x, ml, mu):
-    """
-    Calculates a banded version of the jacobian
+    """ Calculates a banded version of the jacobian
 
     Compatible with the format requested by
-    scipy.integrate.ode
+    :func:`scipy.integrate.ode` (for SciPy >= v0.15).
+
+    Parameters
+    ----------
+    y: array_like of expressions
+    x: array_like of symbols
+    ml: int
+        number of lower bands
+    mu: int
+        number of upper bands
+
+    Returns
+    -------
+    2D array of shape ``(1+ml+mu, len(y))``
     """
     ny = len(y)
     nx = len(x)
@@ -32,6 +52,17 @@ def banded_jacobian(y, x, ml, mu):
 
 
 def check_transforms(fw, bw, symbs):
+    """ Verify validity of a pair of forward and backward transformations
+
+    Parameters
+    ----------
+    fw: expression
+        forward transformation
+    bw: expression
+        backward transformation
+    symbs: iterable of symbols
+        the variables that are transformed
+    """
     for f, b, y in zip(fw, bw, symbs):
         if f.subs(y, b) - y != 0:
             raise ValueError('Incorrect (did you set real=True?) fw: %s'
@@ -42,6 +73,25 @@ def check_transforms(fw, bw, symbs):
 
 
 def transform_exprs_dep(fw, bw, dep_exprs, check=True):
+    """ Transform y[:] in dydx
+
+    Parameters
+    ----------
+    fw: expression
+        forward transformation
+    bw: expression
+        backward transformation
+    dep_exprs: iterable of (symbol, expression) pairs
+        pairs of (dependent variable, derivative expressions),
+        i.e. (y, dydx) pairs
+    check: bool (default: True)
+        whether to verification of the analytic correctness should
+        be performed
+
+    Returns
+    -------
+    List of transformed expressions for dydx
+    """
     if len(fw) != len(dep_exprs) or \
        len(fw) != len(bw):
         raise ValueError("Incompatible lengths")
@@ -53,21 +103,50 @@ def transform_exprs_dep(fw, bw, dep_exprs, check=True):
 
 
 def transform_exprs_indep(fw, bw, dep_exprs, indep, check=True):
+    """ Transform x in dydx
+
+    Parameters
+    ----------
+    fw: expression
+        forward transformation
+    bw: expression
+        backward transformation
+    dep_exprs: iterable of (symbol, expression) pairs
+        pairs of (dependent variable, derivative expressions)
+    check: bool (default: True)
+        whether to verification of the analytic correctness should
+        be performed
+
+    Returns
+    -------
+    List of transformed expressions for dydx
+    """
     if check:
         if fw.subs(indep, bw) - indep != 0:
-                raise ValueError('Incorrect (did you set real=True?) fw: %s'
-                                 % str(fw))
+            fmtstr = 'Incorrect (did you set real=True?) fw: %s'
+            raise ValueError(fmtstr % str(fw))
         if bw.subs(indep, fw) - indep != 0:
-            raise ValueError('Incorrect (did you set real=True?) bw: %s'
-                             % str(bw))
+            fmtstr = 'Incorrect (did you set real=True?) bw: %s'
+            raise ValueError(fmtstr % str(bw))
     dep, exprs = zip(*dep_exprs)
     return [(e/fw.diff(indep)).subs(indep, bw) for e in exprs]
 
 
 def ensure_3args(func):
+    """ Conditionally wrap function to ensure 3 input arguments
+
+    Parameters
+    ----------
+    func: callable
+        with two or three positional arguments
+
+    Returns
+    -------
+    callable which possibly ignores a third positional argument
+    """
     nargs = len(inspect.getargspec(func)[0])
     if nargs == 2:
-        return lambda x, y, params: func(x, y)
+        return lambda x, y, _ignored: func(x, y)
     elif nargs == 3:
         return func
     else: